move firebase to shared
