+++ login/src/presentation/pages/CreateAccount.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { createUserWithEmailAndPassword } from "firebase/auth";
-import { auth } from "../../firebase";
+import { auth } from "shared/firebase";
 
 import {
   Box,
