+++ login/src/presentation/pages/Login.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { signInWithEmailAndPassword } from "firebase/auth";
-import { auth } from "../../firebase";
+import { auth } from "shared/firebase";
 
 import {
   Box,
