+++ shared/package.json
@@ -1,0 +1,9 @@
+
+{
+    "name": "shared",
+    "version": "1.0.0",
+    "type": "module",
+    "dependencies": {
+      "firebase": "^11.10.0"
+    }
+  }
