--- shared/shared.ts
@@ -1,4 +1,0 @@
-export const state = {
-    message: 'Hello from shared'
-  };
-
